Implemented cross-browser cart and fixed lots of stuff Some of the changes: * Implemented rule checking ownership to prevent client-type users from accessing other users models   - missing unauthorized error page * ActiveRecord singleton:   - save() => Properly update entities by implementing update conditions   - where() method now accepts an array and returns all query conditions and params   - Implemented custom query builder, count and sum methods   - Removed find() method and optimized findOne() and findAll() * Re-enabled automatic logout, use session -> logout_time config (seconds, 0 for no logout) * Increased size for decimal values to consistently match the maximum amount of units/total   - Max product quantity set to 999 (hardcoded) * Removed column id for table order_line. Replaced with a composite PK instead * Fixed ajax request not triggering due to data type missmatch * Misc changes such as missing/incorrect return types and indentation * Removed unused Class ActiveQuery. Should probably clean up ActiveRecord and move to ActiveQuery

+++ controllers/ShoppingCartController.php
@@ -2,10 +2,12 @@
 
 namespace controllers;
 
+use framework\App;
 use framework\UserSession;
-use models\Cart;
+use models\BrowserCart;
 use models\Product;
 use models\SignupForm;
+use models\UserCart;
 
 final class ShoppingCartController extends Controller
 {
@@ -59,20 +61,57 @@ final class ShoppingCartController extends Controller
      */
     public function actionIndex(): string
     {
+        if (UserSession::getInstance()->isLoggedIn())
+        {
+            $model = new UserCart();
+        }
+
+        else
+        {
+            $model = new BrowserCart($_COOKIE['items'] ?? []);
+        }
         return $this->render('index', [
-            'model' => new Cart()
+            'model' => $model
         ]);
     }
 
     public function actionAdd(int $id): string
     {
-        if ($id < 1 || !Product::findById($id))
-            return $this->render('error');
-
-        $model = new Cart();
-        $model->addItem($id, $_POST['quantity'] ?? 1);
-
         header('Content-type: application/json');
+
+        if (!$this->getIsAjax())
+            return json_encode([ 'message' => 'This action only supports XMLHttpRequest.' ]);
+
+        if ($id < 1 || !Product::findById($id))
+            return json_encode([ 'message' => 'The product does not exist.' ]);
+
+        if (UserSession::getInstance()->isLoggedIn())
+        {
+            $model = UserCart::findById($id);
+
+            if (!$model)
+            {
+                $model = new UserCart();
+                $model->product_id = $id;
+                $model->quantity = $_POST['quantity'] ?? 1;
+                $model->created_by = App::$user->id;
+            }
+
+            else
+            {
+                $model->quantity += $_POST['quantity'] ?? 1;
+            }
+
+            if ($model->quantity > 0 && $model->quantity < 1000)
+                $model->save();
+        }
+
+        else
+        {
+            $model = new BrowserCart($_COOKIE['items'] ?? []);
+            $model->addItem($id, $_POST['quantity'] ?? 1);
+        }
+
         return json_encode([
             'redirect' => 'reload'
         ]);
@@ -80,23 +119,55 @@ final class ShoppingCartController extends Controller
 
     public function actionUpdate(int $id): string
     {
-        if ($id < 1 || !Product::findById($id))
-            return $this->render('error');
-
-        $model = new Cart();
-
-        $data = $_GET['op'] ?? $_POST['quantity'] ?? false;
-        if ($data && $data === 'up' || $data === 'down')
-        {
-            $model->addItem($id, $data === 'up' ? 1 : -1);
-        }
-
-        else if ($data && $data > 0)
-        {
-            $model->setItemQuantity($id, $data);
-        }
-
         header('Content-type: application/json');
+
+        if (!$this->getIsAjax())
+            return json_encode([ 'message' => 'This action only supports XMLHttpRequest.' ]);
+
+        if ($id < 1 || !Product::findById($id))
+            return json_encode([ 'message' => 'The product does not exist.' ]);
+
+        if (UserSession::getInstance()->isLoggedIn())
+        {
+            $model = UserCart::findById($id);
+
+            if (!$model)
+                return json_encode([ 'message' => 'You cannot update a product that is not in your cart.' ]);
+
+            $data = $_GET['op'] ?? $_POST['quantity'] ?? false;
+            if ($data && $model instanceof UserCart)
+            {
+                if ($data === 'up' || $data === 'down')
+                {
+                    $model->quantity += $data === 'up' ? 1 : -1;
+                }
+
+                else
+                {
+                    $model->quantity = (int) $_POST['quantity'];
+                }
+
+                if ($model->quantity > 0 && $model->quantity < 1000)
+                    $model->save();
+            }
+        }
+
+        else
+        {
+            $model = new BrowserCart($_COOKIE['items'] ?? []);
+
+            $data = $_GET['op'] ?? $_POST['quantity'] ?? false;
+            if ($data && $data === 'up' || $data === 'down')
+            {
+                $model->addItem($id, $data === 'up' ? 1 : -1);
+            }
+
+            else if ($data && $data > 0)
+            {
+                $model->updateItem($id, $data);
+            }
+        }
+
         return json_encode([
             'redirect' => 'redirect'
         ]);
@@ -104,10 +175,31 @@ final class ShoppingCartController extends Controller
 
     public function actionDelete(int $id): string
     {
-        $model = new Cart();
-        $model->deleteItem($id);
-
         header('Content-type: application/json');
+
+        if (!$this->getIsAjax())
+            return json_encode([ 'message' => 'This action only supports XMLHttpRequest.' ]);
+
+        if ($id < 1 || !Product::findById($id))
+            return json_encode([ 'message' => 'The product does not exist.' ]);
+
+        if (UserSession::getInstance()->isLoggedIn())
+        {
+            $model = UserCart::findById($id);
+
+            if (!$model)
+                return json_encode([ 'message' => 'The product is not in your cart.' ]);
+
+            if ($model instanceof UserCart)
+                $model->deleteItem();
+        }
+
+        else
+        {
+            $model = new BrowserCart($_COOKIE['items'] ?? []);
+            $model->deleteItem($id);
+        }
+
         return json_encode([
             'redirect' => 'redirect'
         ]);
@@ -129,8 +221,6 @@ final class ShoppingCartController extends Controller
 
     public function actionThanks(): string
     {
-        (new Cart())->processCart();
-
         return $this->render('thanks');
     }
 }
